Enable ESLint's `class-methods-use-this` rule.
In JavaScript, it's hard to minify a property that is on prototype chain.
Typically, they appears as a pattern as class' instance method.
We cannot remove or mangle a code like `a.foo()` style code.

This rule bans a class instance method
that does not touch any `this` to improve a possibility to minify a code.

After ES Module or CommonJS era (single module per single file),
excluding the case to improve an API ergonomics or requirement to implement an object interface,
we don't have to belong a function that does not touch `this` to a class as like as Java or C++.
Rather, to get a chance to improve a code size performance,
it's better that they should be exported as a standalone function directly.

See:

- https://eslint.org/docs/latest/rules/class-methods-use-this
- https://typescript-eslint.io/rules/class-methods-use-this/