Remove tower_like_ipc.Service.ready()
This instance method was introduced to know whether the target service is ready.

However, this design was poor.
If the service abstract a remote procedure call, this `.ready()`
sometimes tell us that "the remote sever/client is ready now".

But it's a not realtime information. After calling `.ready()`,
if there are something happens,
the sucuceer `.call()` will be failed.
This is not a useful for diagnostics and circit break.

Of cource, their kinds of operations are still useful.
But we think it don't have to be a part of standard interface
which has a resonsibility for RPC abstraction.

It would be better to implement their kinds of operations
into an actual implememtation class and it's more pragmatic approach.

By above reasons, we change the interface in this commit.